Improvements

+++ SceytChatUiKit/src/main/java/com/sceyt/chatuikit/presentation/components/channel_list/channels/viewmodel/ChannelsListViewModelBinding.kt
@@ -5,60 +5,61 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.withResumed
 import com.sceyt.chatuikit.SceytChatUIKit
 import com.sceyt.chatuikit.data.managers.channel.ChannelEventManager
 import com.sceyt.chatuikit.data.models.LoadKeyData
 import com.sceyt.chatuikit.data.models.PaginationResponse
 import com.sceyt.chatuikit.data.models.SceytResponse
 import com.sceyt.chatuikit.data.models.channels.SceytChannel
-import com.sceyt.chatuikit.extensions.customToastSnackBar
 import com.sceyt.chatuikit.extensions.isResumed
 import com.sceyt.chatuikit.logger.SceytLog
 import com.sceyt.chatuikit.persistence.differs.ChannelDiff
 import com.sceyt.chatuikit.persistence.extensions.getPeer
-import com.sceyt.chatuikit.persistence.logicimpl.channel.ChannelUpdateData
 import com.sceyt.chatuikit.persistence.logicimpl.channel.ChannelsCache
 import com.sceyt.chatuikit.presentation.components.channel.header.helpers.TypingCancelHelper
 import com.sceyt.chatuikit.presentation.components.channel_list.channels.ChannelListView
 import com.sceyt.chatuikit.presentation.components.channel_list.channels.adapter.ChannelListItem
 import com.sceyt.chatuikit.presentation.components.channel_list.channels.adapter.ChannelListItem.ChannelItem
-import com.sceyt.chatuikit.presentation.components.channel_list.channels.adapter.ChannelsComparatorDescBy
+import com.sceyt.chatuikit.presentation.components.channel_list.channels.adapter.ChannelsItemComparatorBy
 import com.sceyt.chatuikit.presentation.components.channel_list.search.SearchChannelInputView
 import com.sceyt.chatuikit.services.SceytPresenceChecker
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.withContext
-import java.util.concurrent.ConcurrentHashMap
+
 
 fun ChannelsViewModel.bind(channelListView: ChannelListView, lifecycleOwner: LifecycleOwner) {
 
     val typingCancelHelper by lazy { TypingCancelHelper() }
-    val needToUpdateChannelsAfterResume = ConcurrentHashMap<Long, ChannelUpdateData>()
-    val newAddedChannelJobs = ConcurrentHashMap<Long, Job>()
+    var needSubmitOnResume: List<ChannelListItem>? = null
+    val mutexUpdateList = Mutex()
+
+    fun getUpdateAfterOnResumeData(): List<ChannelListItem> {
+        return (needSubmitOnResume.takeIf { !it.isNullOrEmpty() }
+                ?: channelListView.getData()) ?: mutableListOf()
+    }
 
     getChannels(0, query = searchQuery)
 
     viewModelScope.launch {
         lifecycleOwner.repeatOnLifecycle(Lifecycle.State.RESUMED) {
-            channelListView.post {
-                if (needToUpdateChannelsAfterResume.isNotEmpty()) {
-                    val isCanceled = channelListView.cancelLastSort()
-                    var needSort = isCanceled
-                    needToUpdateChannelsAfterResume.values.forEach { data ->
-                        val diff = channelListView.channelUpdated(data.channel)
-                        if (diff != null && !needSort) {
-                            if (diff.lastMessageChanged || data.needSorting)
-                                needSort = true
+            if (needSubmitOnResume.isNullOrEmpty())
+                return@repeatOnLifecycle
+
+            lifecycleOwner.lifecycleScope.launch(Dispatchers.Default) {
+                mutexUpdateList.withLock {
+                    val newData = needSubmitOnResume?.sortedWith(ChannelsItemComparatorBy(config.order))
+                    needSubmitOnResume = null
+                    if (newData != null) {
+                        withContext(Dispatchers.Main) {
+                            channelListView.setChannelsList(newData)
                         }
                     }
-                    needToUpdateChannelsAfterResume.clear()
-                    if (needSort)
-                        channelListView.sortChannelsBy(config.order)
                 }
             }
         }
@@ -97,41 +98,40 @@ fun ChannelsViewModel.bind(channelListView: ChannelListView, lifecycleOwner: Lif
     loadChannelsFlow.onEach(::initChannelsResponse).launchIn(lifecycleOwner.lifecycleScope)
 
     ChannelsCache.channelsDeletedFlow.onEach { channelIds ->
-        lifecycleOwner.lifecycleScope.launch {
+        if (!lifecycleOwner.isResumed()) {
+            lifecycleOwner.lifecycleScope.launch(Dispatchers.Default) {
+                mutexUpdateList.withLock {
+                    needSubmitOnResume = getUpdateAfterOnResumeData().filter {
+                        (it as? ChannelItem)?.channel?.id !in channelIds
+                    }
+                }
+            }
+        } else {
             channelIds.forEach { channelId ->
-                newAddedChannelJobs[channelId]?.apply {
-                    cancel()
-                    newAddedChannelJobs.remove(channelId)
-                }
-                needToUpdateChannelsAfterResume.remove(channelId)
-                lifecycleOwner.withResumed {
-                    channelListView.deleteChannel(channelId, searchQuery)
-                }
+                channelListView.deleteChannel(channelId, searchQuery)
             }
         }
     }.launchIn(viewModelScope)
 
     ChannelsCache.channelUpdatedFlow.onEach { data ->
         if (!lifecycleOwner.isResumed()) {
-            val old = needToUpdateChannelsAfterResume[data.channel.id]
-            // Check maybe already exists with needSorting
-            val newData = if (old != null && old.needSorting) {
-                data.copy(needSorting = true)
-            } else data
-            needToUpdateChannelsAfterResume[data.channel.id] = newData
-        } else {
-            lifecycleOwner.lifecycleScope.launch {
-                val isCanceled = channelListView.cancelLastSort()
-                val diff = channelListView.channelUpdated(data.channel)
-                if (diff != null) {
-                    if (diff.lastMessageChanged || data.needSorting || isCanceled)
-                        channelListView.sortChannelsBy(SceytChatUIKit.config.channelListOrder)
-                    SceytLog.i("ChannelsCache", "viewModel: id: ${data.channel.id}  body: ${data.channel.lastMessage?.body} draft:${data.channel.draftMessage?.body}  unreadCount ${data.channel.newMessageCount}" +
-                            " isResumed ${lifecycleOwner.isResumed()} hasDifference: ${diff.hasDifference()} lastMessageChanged: ${diff.lastMessageChanged} needSorting: ${data.needSorting}")
-                } else {
-                    SceytLog.i("ChannelsCache", "viewModel: id: ${data.channel.id}  body: ${data.channel.lastMessage?.body}  unreadCount ${data.channel.newMessageCount}" +
-                            " isResumed ${lifecycleOwner.isResumed()} but started getChannels ")
+            lifecycleOwner.lifecycleScope.launch(Dispatchers.Default) {
+                mutexUpdateList.withLock {
+                    needSubmitOnResume = getUpdateAfterOnResumeData().map {
+                        if (it is ChannelItem && it.channel.id == data.channel.id) {
+                            it.copy(channel = data.channel)
+                        } else it
+                    }
                 }
+            }
+        } else {
+            val isCanceled = channelListView.cancelLastSort()
+            channelListView.channelUpdated(data.channel) {
+                if (data.diff.lastMessageChanged || data.needSorting || isCanceled)
+                    channelListView.sortChannelsBy(SceytChatUIKit.config.channelListOrder)
+
+                SceytLog.i("ChannelsCache", "viewModel: id: ${data.channel.id}  body: ${data.channel.lastMessage?.body} draft:${data.channel.draftMessage?.body}  unreadCount ${data.channel.newMessageCount}" +
+                        " isResumed ${lifecycleOwner.isResumed()} hasDifference: ${data.diff.hasDifference()} lastMessageChanged: ${data.diff.lastMessageChanged} needSorting: ${data.needSorting}")
             }
         }
     }.launchIn(viewModelScope)
@@ -142,106 +142,79 @@ fun ChannelsViewModel.bind(channelListView: ChannelListView, lifecycleOwner: Lif
         }
     }.launchIn(viewModelScope)
 
-    fun createJobToAddNewChannelWithOnResumed(sceytChannel: SceytChannel) {
-        val job = viewModelScope.launch {
-            lifecycleOwner.withResumed {
-                val updatedChannel = needToUpdateChannelsAfterResume[sceytChannel.id]?.channel
-                        ?: sceytChannel
-                channelListView.cancelLastSort()
-                channelListView.addNewChannelAndSort(config.order, ChannelItem(updatedChannel))
-                newAddedChannelJobs.remove(sceytChannel.id)
-            }
-        }
-        newAddedChannelJobs[sceytChannel.id] = job
-    }
-
     ChannelsCache.channelAddedFlow
         .filter { config.isValidForConfig(it) }
-        .onEach(::createJobToAddNewChannelWithOnResumed)
+        .onEach {
+            if (!lifecycleOwner.isResumed()) {
+                lifecycleOwner.lifecycleScope.launch(Dispatchers.Default) {
+                    mutexUpdateList.withLock {
+                        val newData = getUpdateAfterOnResumeData().toMutableList()
+                        if (newData.none { item -> (item as? ChannelItem)?.channel?.id == it.id }) {
+                            newData.add(ChannelItem(it))
+                            needSubmitOnResume = newData
+                        }
+                    }
+                }
+            } else {
+                channelListView.addNewChannelAndSort(config.order, ChannelItem(it))
+            }
+        }
         .launchIn(viewModelScope)
 
     ChannelsCache.pendingChannelCreatedFlow
         .filter { config.isValidForConfig(it.second) }
         .onEach { (pendingChannelId, newChannel) ->
-            channelListView.replaceChannel(pendingChannelId, newChannel)
             if (!lifecycleOwner.isResumed()) {
-                newAddedChannelJobs[pendingChannelId]?.let {
-                    it.cancel()
-                    newAddedChannelJobs.remove(pendingChannelId)
+                lifecycleOwner.lifecycleScope.launch(Dispatchers.Default) {
+                    mutexUpdateList.withLock {
+                        val newData = getUpdateAfterOnResumeData().toMutableList()
+                        newData.removeAll { (it as? ChannelItem)?.channel?.id == pendingChannelId }
+                        if (newData.none { (it as? ChannelItem)?.channel?.id == newChannel.id }) {
+                            newData.add(ChannelItem(newChannel))
+                            needSubmitOnResume = newData
+                        }
+                    }
                 }
-                createJobToAddNewChannelWithOnResumed(newChannel)
-            }
+            } else
+                channelListView.replaceChannel(pendingChannelId, newChannel)
         }.launchIn(viewModelScope)
 
     ChannelsCache.channelDraftMessageChangesFlow.onEach { channel ->
-        channelListView.channelUpdatedWithDiff(channel, ChannelDiff.DEFAULT_FALSE.copy(lastMessageChanged = true))
         if (!lifecycleOwner.isResumed()) {
-            val pendingUpdate = needToUpdateChannelsAfterResume[channel.id]
-            if (pendingUpdate != null) {
-                pendingUpdate.channel = channel
-            } else
-                needToUpdateChannelsAfterResume[channel.id] = ChannelUpdateData(channel, false)
+            lifecycleOwner.lifecycleScope.launch {
+                mutexUpdateList.withLock {
+                    needSubmitOnResume = getUpdateAfterOnResumeData().map {
+                        if (it is ChannelItem && it.channel.id == channel.id) {
+                            it.copy(channel = channel)
+                        } else it
+                    }
+                }
+            }
+        } else {
+            channelListView.channelUpdatedWithDiff(channel, ChannelDiff.DEFAULT_FALSE.copy(lastMessageChanged = true))
         }
     }.launchIn(viewModelScope)
 
     ChannelsCache.newChannelsOnSync
         .onEach { (_, channels) ->
-            lifecycleOwner.withResumed {
-                lifecycleOwner.lifecycleScope.launch(Dispatchers.Default) {
-                    // Filter channels by config
-                    val filtered = channels.filter { config.isValidForConfig(it) }
-                    if (filtered.isNotEmpty()) {
-                        val loadedChannels = channelListView.getData()?.mapNotNull {
+            lifecycleOwner.lifecycleScope.launch {
+                if (lifecycleOwner.isResumed()) {
+                    val loadedChannels = channelListView.getData()?.mapNotNull {
+                        (it as? ChannelItem)?.channel
+                    } ?: emptyList()
+                    val dataToSubmit = initDataOnNewChannelsOnSync(loadedChannels, channels)
+                            ?: return@launch
+                    channelListView.setChannelsList(dataToSubmit)
+                } else {
+                    mutexUpdateList.withLock {
+                        val data = getUpdateAfterOnResumeData().mapNotNull {
                             (it as? ChannelItem)?.channel
-                        }?.toMutableSet() ?: mutableSetOf()
-
-                        // If loadedChannels are empty and not loading data from server, it means we can setData,
-                        // otherwise we filter only channels which are between loaded channels and
-                        // insert them to the list.
-                        if (loadedChannels.isEmpty()) {
-                            if (loadingFromServer || loadingFromDb) return@launch
-                            val sorted = filtered.sortedWith(ChannelsComparatorDescBy(config.order))
-                            val date = mapToChannelItem(data = sorted, hasNext = false)
-                            SceytLog.i("syncResultUpdate", "loaded channels are empty, set data : ${sorted.map { it.channelSubject }}")
-                            withContext(Dispatchers.Main) {
-                                channelListView.setChannelsList(date)
-                            }
-                        } else {
-                            // Get last channel to understand where to insert new channels
-                            val lastChannel = loadedChannels.last()
-                            val sorted = filtered.toSet().plus(lastChannel).sortedWith(ChannelsComparatorDescBy(config.order))
-                            val index = sorted.indexOf(lastChannel)
-
-                            // If index is last and we have more channels, we don't need to insert them,
-                            // because they will be inserted by next page loading
-                            if (index == loadedChannels.size - 1 && (hasNext || hasNextDb)) {
-                                return@launch
-                            }
-                            // Get channels which need to be inserted
-                            sorted.subList(0, index).forEach {
-                                loadedChannels.add(it)
-                            }
-                            var newData: List<ChannelListItem> = loadedChannels.sortedWith(ChannelsComparatorDescBy(config.order)).map {
-                                ChannelItem(it)
-                            }
-
-                            if (hasNext || hasNextDb)
-                                newData = newData.plus(ChannelListItem.LoadingMoreItem)
-
-                            SceytLog.i("syncResultUpdate", "apply synced channels : ${
-                                newData.map {
-                                    (it as? ChannelItem)?.channel?.channelSubject ?: it.toString()
-                                }
-                            }")
-                            withContext(Dispatchers.Main) {
-                                channelListView.setChannelsList(newData)
-                            }
                         }
+                        needSubmitOnResume = initDataOnNewChannelsOnSync(data, channels)
                     }
                 }
             }
-        }
-        .launchIn(viewModelScope)
+        }.launchIn(viewModelScope)
 
     ChannelEventManager.onChannelTypingEventFlow
         .filter { it.member.id != SceytChatUIKit.chatUIFacade.myId }
@@ -251,16 +224,6 @@ fun ChannelsViewModel.bind(channelListView: ChannelListView, lifecycleOwner: Lif
             }
             channelListView.onTyping(it)
         }.launchIn(lifecycleOwner.lifecycleScope)
-
-    blockUserLiveData.observe(lifecycleOwner) {
-        when (it) {
-            is SceytResponse.Success -> {
-                channelListView.userBlocked(it.data)
-            }
-
-            is SceytResponse.Error -> customToastSnackBar(channelListView, it.message ?: "")
-        }
-    }
 
     pageStateLiveData.observe(lifecycleOwner) {
         channelListView.updateStateView(it)
